+++ internal/cmd/company/iam/remove.go
@@ -17,6 +17,7 @@ package iam
 
 import (
 	"github.com/mia-platform/miactl/internal/clioptions"
+	"github.com/mia-platform/miactl/internal/cmd/company/iam/group"
 	"github.com/mia-platform/miactl/internal/cmd/company/iam/user"
 	"github.com/spf13/cobra"
 )
@@ -31,6 +32,7 @@ from the company selected via the flag or context`,
 
 	cmd.AddCommand(
 		user.RemoveCmd(options),
+		group.RemoveCmd(options),
 	)
 
 	return cmd